backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
pymongo==4.12.1
python-dotenv==1.1.0
PyYAML==6.0.2
gunicorn==23.0.0
